# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryContext.java
#	modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryEnumObjectImpl.java
#	modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryObjectImpl.java
#	modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryObjectOffheapImpl.java
#	modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryUtils.java
#	modules/core/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedMarshaller.java
#	modules/core/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedMarshallerUtils.java
#	modules/core/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedObjectInputStream.java
#	modules/core/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedObjectStreamRegistry.java
#	modules/core/src/main/java/org/apache/ignite/internal/util/IgniteUtils.java
#	modules/core/src/main/java/org/apache/ignite/marshaller/jdk/JdkMarshallerObjectInputStream.java
#	modules/core/src/test/java/org/apache/ignite/internal/marshaller/optimized/OptimizedObjectStreamSelfTest.java
#	modules/core/src/test/java/org/apache/ignite/testsuites/IgniteBinaryObjectsCacheTestSuite3.java

+++ modules/cassandra/store/src/test/java/org/apache/ignite/tests/load/PersonGenerator.java
@@ -30,7 +30,7 @@ public class PersonGenerator implements Generator {
     private static final Date DATE = new Date();
 
     /** */
-    private static final List<String> PHONES = new LinkedList<String>(){{
+    private static final List<String> PHONES = new LinkedList<String>() {{
         add("1234567");
         add("7654321");
         add("1289054");
@@ -38,6 +38,6 @@ public class PersonGenerator implements Generator {
 
     /** {@inheritDoc} */
     @Override public Object generate(long i) {
-        return new Person(i, Long.toString(i), Long.toString(i), (int)(i % 100), i % 2 == 0, i, i, DATE, PHONES);
+        return new Person(i, Long.toString(i), Long.toString(i), (short)(i % 100), i % 2 == 0, i, i, DATE, PHONES);
     }
 }
